etls/facebook/requirements.txt
alembic==1.4.3
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
requests==2.24.0
six==1.15.0
urllib3==1.25.11
